fix(ops): move sub-nav inside content area — was rendering above main nav
- Extracted OpsSubNav component from layout.tsx to src/components/ops/OpsSubNav.tsx
- Layout.tsx now passthrough (just renders children)
- OpsSubNav rendered inside AppLayout in both ops/page.tsx and ops/bookkeeping/page.tsx
- Sub-nav appears below main nav and module tabs, above page content
- No existing functionality modified
